:truck: Reorganize project files
