--- lib/connect_four/exceptions.rb
@@ -1,5 +1,0 @@
-class PlayOutOfTurnError < StandardError
-end
-
-class FullColumnError < StandardError
-end
