+++ lib/dice.rb
