+++ lib/exceptions.rb
@@ -1,0 +1,4 @@
+module Exceptions
+    class PlayOutOfTurnError < StandardError; end
+    class FullColumnError < StandardError; end
+end
