+++ test/models/test_connect_four.rb
@@ -2,7 +2,7 @@ require "minitest/autorun"
 require "minitest/reporters"
 Minitest::Reporters.use! Minitest::Reporters::ProgressReporter.new
 
-require_relative "../../lib/connect_four/connect_four"
+require_relative "../../app/models/connect_four"
 
 class ConnectFourTest < Minitest::Test
     def setup
@@ -48,13 +48,13 @@ class ConnectFourTest < Minitest::Test
 
         connect4_full_col = ConnectFour.new(full_col)
 
-        assert_raises FullColumnError do
+        assert_raises Exceptions::FullColumnError do
             connect4_full_col.play(1,1)
         end
     end
 
     def test_play_should_raise_when_user_plays_out_of_turn
-        assert_raises PlayOutOfTurnError do
+        assert_raises Exceptions::PlayOutOfTurnError do
             @connect4.play(2, 1)
         end
     end
